+++ frontend/plugins/mongolian_ui/src/modules/erkhet-sync/settings/pipeline-remainder-config/components/PipelineRemainderConfigAddSheet.tsx
@@ -3,9 +3,9 @@ import { Button, Form, Input, Sheet } from 'erxes-ui';
 import { useState } from 'react';
 import { useForm } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
-import { SelectBoard, SelectPipeline, SelectStage } from 'ui-modules';
 import { addPipelineRemainderConfigSchema } from '../constants/addPipelineRemainderConfigSchema';
 import { AddPipelineRemainderConfig } from '../types';
+import { PipelineSelectorFields } from '../../shared/components/PipelineSelectorFields';
 
 const defaultValues: AddPipelineRemainderConfig = {
   title: '',
@@ -28,9 +28,6 @@ export const PipelineRemainderConfigAddSheet = ({ onSubmit, loading }: Props) =>
     resolver: zodResolver(addPipelineRemainderConfigSchema),
     defaultValues,
   });
-
-  const selectedBoardId = form.watch('boardId');
-  const selectedPipelineId = form.watch('pipelineId');
 
   const handleSubmit = async (data: AddPipelineRemainderConfig) => {
     await onSubmit(data);
@@ -74,26 +71,6 @@ export const PipelineRemainderConfigAddSheet = ({ onSubmit, loading }: Props) =>
                       )}
                     />
                     <Form.Field
-                      control={form.control}
-                      name="pipelineId"
-                      render={({ field }) => (
-                        <Form.Item>
-                          <Form.Label>Pipeline</Form.Label>
-                          <SelectPipeline
-                            mode="single"
-                            value={field.value}
-                            onValueChange={(value) => {
-                              field.onChange(value as string);
-                              form.setValue('stageId', '');
-                            }}
-                            boardId={selectedBoardId || undefined}
-                            placeholder="Select pipeline"
-                          />
-                          <Form.Message />
-                        </Form.Item>
-                      )}
-                    />
-                    <Form.Field
                       name="account"
                       control={form.control}
                       render={({ field }) => (
@@ -102,45 +79,6 @@ export const PipelineRemainderConfigAddSheet = ({ onSubmit, loading }: Props) =>
                           <Form.Control>
                             <Input {...field} placeholder="Account" />
                           </Form.Control>
-                          <Form.Message />
-                        </Form.Item>
-                      )}
-                    />
-                  </div>
-                  <div className="flex flex-col gap-4">
-                    <Form.Field
-                      control={form.control}
-                      name="boardId"
-                      render={({ field }) => (
-                        <Form.Item>
-                          <Form.Label>Board</Form.Label>
-                          <SelectBoard
-                            mode="single"
-                            value={field.value}
-                            onValueChange={(value) => {
-                              field.onChange(value as string);
-                              form.setValue('pipelineId', '');
-                              form.setValue('stageId', '');
-                            }}
-                            placeholder="Select board"
-                          />
-                          <Form.Message />
-                        </Form.Item>
-                      )}
-                    />
-                    <Form.Field
-                      control={form.control}
-                      name="stageId"
-                      render={({ field }) => (
-                        <Form.Item>
-                          <Form.Label>Stage</Form.Label>
-                          <SelectStage
-                            mode="single"
-                            value={field.value}
-                            onValueChange={(value) => field.onChange(value as string)}
-                            pipelineId={selectedPipelineId || undefined}
-                            placeholder="Select stage"
-                          />
                           <Form.Message />
                         </Form.Item>
                       )}
@@ -158,6 +96,9 @@ export const PipelineRemainderConfigAddSheet = ({ onSubmit, loading }: Props) =>
                         </Form.Item>
                       )}
                     />
+                  </div>
+                  <div className="flex flex-col gap-4">
+                    <PipelineSelectorFields form={form} />
                   </div>
                 </div>
               </div>
