+++ frontend/plugins/mongolian_ui/src/modules/erkhet-sync/settings/pipeline-remainder-config/components/PipelineRemainderConfigEditSheet.tsx
@@ -2,10 +2,10 @@ import { Button, Form, Input, Sheet } from 'erxes-ui';
 import { useEffect } from 'react';
 import { useForm } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
-import { SelectBoard, SelectPipeline, SelectStage } from 'ui-modules';
 import { addPipelineRemainderConfigSchema } from '../constants/addPipelineRemainderConfigSchema';
 import { AddPipelineRemainderConfig } from '../types';
 import { TRemainderConfigRow } from '../hooks/usePipelineRemainderConfigs';
+import { PipelineSelectorFields } from '../../shared/components/PipelineSelectorFields';
 
 interface Props {
   config: TRemainderConfigRow;
@@ -45,9 +45,6 @@ export const PipelineRemainderConfigEditSheet = ({
     });
   }, [config, form]);
 
-  const selectedBoardId = form.watch('boardId');
-  const selectedPipelineId = form.watch('pipelineId');
-
   const handleSubmit = async (data: AddPipelineRemainderConfig) => {
     await onSubmit(config._id, data);
     onOpenChange(false);
@@ -83,26 +80,6 @@ export const PipelineRemainderConfigEditSheet = ({
                       )}
                     />
                     <Form.Field
-                      control={form.control}
-                      name="pipelineId"
-                      render={({ field }) => (
-                        <Form.Item>
-                          <Form.Label>Pipeline</Form.Label>
-                          <SelectPipeline
-                            mode="single"
-                            value={field.value}
-                            onValueChange={(value) => {
-                              field.onChange(value as string);
-                              form.setValue('stageId', '');
-                            }}
-                            boardId={selectedBoardId || undefined}
-                            placeholder="Select pipeline"
-                          />
-                          <Form.Message />
-                        </Form.Item>
-                      )}
-                    />
-                    <Form.Field
                       name="account"
                       control={form.control}
                       render={({ field }) => (
@@ -111,45 +88,6 @@ export const PipelineRemainderConfigEditSheet = ({
                           <Form.Control>
                             <Input {...field} placeholder="Account" />
                           </Form.Control>
-                          <Form.Message />
-                        </Form.Item>
-                      )}
-                    />
-                  </div>
-                  <div className="flex flex-col gap-4">
-                    <Form.Field
-                      control={form.control}
-                      name="boardId"
-                      render={({ field }) => (
-                        <Form.Item>
-                          <Form.Label>Board</Form.Label>
-                          <SelectBoard
-                            mode="single"
-                            value={field.value}
-                            onValueChange={(value) => {
-                              field.onChange(value as string);
-                              form.setValue('pipelineId', '');
-                              form.setValue('stageId', '');
-                            }}
-                            placeholder="Select board"
-                          />
-                          <Form.Message />
-                        </Form.Item>
-                      )}
-                    />
-                    <Form.Field
-                      control={form.control}
-                      name="stageId"
-                      render={({ field }) => (
-                        <Form.Item>
-                          <Form.Label>Stage</Form.Label>
-                          <SelectStage
-                            mode="single"
-                            value={field.value}
-                            onValueChange={(value) => field.onChange(value as string)}
-                            pipelineId={selectedPipelineId || undefined}
-                            placeholder="Select stage"
-                          />
                           <Form.Message />
                         </Form.Item>
                       )}
@@ -167,6 +105,9 @@ export const PipelineRemainderConfigEditSheet = ({
                         </Form.Item>
                       )}
                     />
+                  </div>
+                  <div className="flex flex-col gap-4">
+                    <PipelineSelectorFields form={form} />
                   </div>
                 </div>
               </div>
