+++ frontend/plugins/mongolian_ui/src/modules/erkhet-sync/settings/stage-in-erkhet-movement-config/components/MovementConfigColumns.tsx
@@ -1,133 +1,12 @@
-import {
-  IconEdit,
-  IconAlignLeft,
-  IconAt,
-  IconLayoutKanban,
-  IconTrash,
-  IconUser,
-} from '@tabler/icons-react';
+import { IconAlignLeft, IconAt, IconLayoutKanban, IconUser } from '@tabler/icons-react';
 import { CellContext, ColumnDef } from '@tanstack/react-table';
-import {
-  Button,
-  Combobox,
-  Command,
-  Popover,
-  RecordTable,
-  RecordTableInlineCell,
-  useConfirm,
-} from 'erxes-ui';
+import { RecordTable, RecordTableInlineCell } from 'erxes-ui';
 import { checkboxColumn } from 'erxes-ui/modules/record-table/components/CheckboxColumn';
-import { useState } from 'react';
 import { TMovementErkhetConfig } from '../types';
 import { MovementConfigEditSheet } from './MovementConfigEditSheet';
+import { ErkhetConfigTitleCell, ErkhetConfigMoreCell } from '../../shared/components/ErkhetConfigColumnCells';
 
 type TConfigRow = TMovementErkhetConfig & { _id: string };
-
-const TitleCell = ({
-  config,
-  onEdit,
-  editLoading,
-}: {
-  config: TConfigRow;
-  onEdit: (id: string, data: TMovementErkhetConfig) => Promise<void>;
-  editLoading: boolean;
-}) => {
-  const [open, setOpen] = useState(false);
-
-  return (
-    <>
-      <RecordTableInlineCell
-        className="cursor-pointer font-medium"
-        onClick={() => setOpen(true)}
-      >
-        {config.title || '—'}
-      </RecordTableInlineCell>
-      {open && (
-        <MovementConfigEditSheet
-          config={config}
-          open={open}
-          onOpenChange={setOpen}
-          onSubmit={onEdit}
-          loading={editLoading}
-        />
-      )}
-    </>
-  );
-};
-
-const MovementMoreColumnCell = ({
-  cell,
-  onEdit,
-  onDelete,
-  editLoading,
-}: {
-  cell: CellContext<TConfigRow, unknown>;
-  onEdit: (id: string, data: TMovementErkhetConfig) => Promise<void>;
-  onDelete: (id: string) => void;
-  editLoading: boolean;
-}) => {
-  const [editOpen, setEditOpen] = useState(false);
-  const config = cell.row.original;
-  const { confirm } = useConfirm();
-
-  const handleDelete = () => {
-    confirm({ message: 'Are you sure you want to delete this config?' }).then(
-      () => onDelete(config._id),
-    );
-  };
-
-  return (
-    <>
-      <Popover>
-        <Popover.Trigger asChild>
-          <RecordTable.MoreButton className="w-full h-full" />
-        </Popover.Trigger>
-        <Combobox.Content
-          align="start"
-          className="w-[220px] min-w-0 [&>button]:cursor-pointer"
-          onClick={(e) => e.stopPropagation()}
-        >
-          <Command>
-            <Command.List>
-              <Command.Item asChild>
-                <Button
-                  variant="ghost"
-                  size="sm"
-                  className="w-full justify-start h-8"
-                  onClick={() => setEditOpen(true)}
-                >
-                  <IconEdit className="size-4" />
-                  Edit
-                </Button>
-              </Command.Item>
-              <Command.Item asChild>
-                <Button
-                  variant="ghost"
-                  size="sm"
-                  className="w-full justify-start h-8 text-destructive hover:text-destructive"
-                  onClick={handleDelete}
-                  disabled={editLoading}
-                >
-                  <IconTrash className="size-4" />
-                  Delete
-                </Button>
-              </Command.Item>
-            </Command.List>
-          </Command>
-        </Combobox.Content>
-      </Popover>
-      {editOpen && (
-        <MovementConfigEditSheet
-          config={config}
-          open={editOpen}
-          onOpenChange={setEditOpen}
-          onSubmit={onEdit}
-          loading={editLoading}
-        />
-      )}
-    </>
-  );
-};
 
 export const buildMovementConfigColumns = (
   onEdit: (id: string, data: TMovementErkhetConfig) => Promise<void>,
@@ -136,12 +15,20 @@ export const buildMovementConfigColumns = (
 ): ColumnDef<TConfigRow>[] => [
   {
     id: 'more',
-    cell: (cell) => (
-      <MovementMoreColumnCell
+    cell: (cell: CellContext<TConfigRow, unknown>) => (
+      <ErkhetConfigMoreCell
         cell={cell}
-        onEdit={onEdit}
         onDelete={onDelete}
         editLoading={editLoading}
+        renderEditSheet={(open, onOpenChange) => (
+          <MovementConfigEditSheet
+            config={cell.row.original}
+            open={open}
+            onOpenChange={onOpenChange}
+            onSubmit={onEdit}
+            loading={editLoading}
+          />
+        )}
       />
     ),
     size: 25,
@@ -152,10 +39,17 @@ export const buildMovementConfigColumns = (
     accessorKey: 'title',
     header: () => <RecordTable.InlineHead icon={IconAlignLeft} label="Title" />,
     cell: ({ row }) => (
-      <TitleCell
+      <ErkhetConfigTitleCell
         config={row.original}
-        onEdit={onEdit}
-        editLoading={editLoading}
+        renderEditSheet={(open, onOpenChange) => (
+          <MovementConfigEditSheet
+            config={row.original}
+            open={open}
+            onOpenChange={onOpenChange}
+            onSubmit={onEdit}
+            loading={editLoading}
+          />
+        )}
       />
     ),
     size: 200,
@@ -165,35 +59,25 @@ export const buildMovementConfigColumns = (
     accessorKey: 'userEmail',
     header: () => <RecordTable.InlineHead icon={IconAt} label="User Email" />,
     cell: ({ cell }) => (
-      <RecordTableInlineCell>
-        {(cell.getValue() as string) || '—'}
-      </RecordTableInlineCell>
+      <RecordTableInlineCell>{(cell.getValue() as string) || '—'}</RecordTableInlineCell>
     ),
     size: 200,
   },
   {
     id: 'defaultCustomer',
     accessorKey: 'defaultCustomer',
-    header: () => (
-      <RecordTable.InlineHead icon={IconUser} label="Default Customer" />
-    ),
+    header: () => <RecordTable.InlineHead icon={IconUser} label="Default Customer" />,
     cell: ({ cell }) => (
-      <RecordTableInlineCell>
-        {(cell.getValue() as string) || '—'}
-      </RecordTableInlineCell>
+      <RecordTableInlineCell>{(cell.getValue() as string) || '—'}</RecordTableInlineCell>
     ),
     size: 160,
   },
   {
     id: 'chooseResponseField',
     accessorKey: 'chooseResponseField',
-    header: () => (
-      <RecordTable.InlineHead icon={IconLayoutKanban} label="Response Field" />
-    ),
+    header: () => <RecordTable.InlineHead icon={IconLayoutKanban} label="Response Field" />,
     cell: ({ cell }) => (
-      <RecordTableInlineCell>
-        {(cell.getValue() as string) || '—'}
-      </RecordTableInlineCell>
+      <RecordTableInlineCell>{(cell.getValue() as string) || '—'}</RecordTableInlineCell>
     ),
     size: 160,
   },
