+++ frontend/plugins/mongolian_ui/src/modules/erkhet-sync/settings/stage-in-erkhet-movement-config/components/MovementConfigRecordTable.tsx
@@ -1,8 +1,7 @@
-import { IconSettings2 } from '@tabler/icons-react';
-import { RecordTable } from 'erxes-ui';
 import { TMovementErkhetConfig } from '../types';
 import { buildMovementConfigColumns } from './MovementConfigColumns';
 import { MovementConfigCommandBar } from './MovementConfigCommandBar';
+import { ErkhetConfigRecordTable } from '../../shared/components/ErkhetConfigRecordTable';
 
 type TConfigRow = TMovementErkhetConfig & { _id: string };
 
@@ -20,36 +19,11 @@ export const MovementConfigRecordTable = ({
   onDelete,
   onDeleteMany,
   editLoading,
-}: Props) => {
-  const columns = buildMovementConfigColumns(onEdit, onDelete, editLoading);
-
-  return (
-    <RecordTable.Provider
-      columns={columns}
-      data={configs}
-      className="m-3"
-      tableOptions={{ enableRowSelection: true } as any}
-    >
-      <RecordTable>
-        <RecordTable.Header />
-        <RecordTable.Body>
-          <RecordTable.RowList />
-        </RecordTable.Body>
-      </RecordTable>
-      {configs.length === 0 && (
-        <div className="flex items-center justify-center py-20">
-          <div className="flex flex-col items-center gap-3 text-center">
-            <IconSettings2 size={48} className="text-gray-400" />
-            <h3 className="text-lg font-semibold text-gray-900">
-              No configs yet
-            </h3>
-            <p className="text-sm text-gray-500">
-              Create your first movement config using the button above.
-            </p>
-          </div>
-        </div>
-      )}
-      <MovementConfigCommandBar onDeleteMany={onDeleteMany} loading={editLoading} />
-    </RecordTable.Provider>
-  );
-};
+}: Props) => (
+  <ErkhetConfigRecordTable
+    configs={configs}
+    columns={buildMovementConfigColumns(onEdit, onDelete, editLoading)}
+    emptyDescription="Create your first movement config using the button above."
+    commandBar={<MovementConfigCommandBar onDeleteMany={onDeleteMany} loading={editLoading} />}
+  />
+);
