+++ frontend/plugins/mongolian_ui/src/modules/erkhet-sync/settings/stage-in-return-erkhet-config/components/ReturnErkhetConfigAddSheet.tsx
@@ -3,10 +3,10 @@ import { Button, Form, Input, Select, Sheet } from 'erxes-ui';
 import { useState } from 'react';
 import { useForm } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
-import { SelectBoard, SelectPipeline, SelectStage } from 'ui-modules';
 import { addStageInReturnErkhetConfigSchema } from '../constants/addStageInReturnErkhetConfigSchema';
 import { RETURN_TYPES } from '../constants/returnTypesData';
 import { TReturnErkhetConfig } from '../types';
+import { PipelineSelectorFields } from '../../shared/components/PipelineSelectorFields';
 
 const defaultValues: TReturnErkhetConfig = {
   title: '',
@@ -29,9 +29,6 @@ export const ReturnErkhetConfigAddSheet = ({ onSubmit, loading }: Props) => {
     resolver: zodResolver(addStageInReturnErkhetConfigSchema),
     defaultValues,
   });
-
-  const selectedBoardId = form.watch('boardId');
-  const selectedPipelineId = form.watch('pipelineId');
 
   const handleSubmit = async (data: TReturnErkhetConfig) => {
     await onSubmit(data);
@@ -75,26 +72,6 @@ export const ReturnErkhetConfigAddSheet = ({ onSubmit, loading }: Props) => {
                       )}
                     />
                     <Form.Field
-                      control={form.control}
-                      name="pipelineId"
-                      render={({ field }) => (
-                        <Form.Item>
-                          <Form.Label>Pipeline</Form.Label>
-                          <SelectPipeline
-                            mode="single"
-                            value={field.value}
-                            onValueChange={(value) => {
-                              field.onChange(value as string);
-                              form.setValue('stageId', '');
-                            }}
-                            boardId={selectedBoardId || undefined}
-                            placeholder="Select pipeline"
-                          />
-                          <Form.Message />
-                        </Form.Item>
-                      )}
-                    />
-                    <Form.Field
                       name="userEmail"
                       control={form.control}
                       render={({ field }) => (
@@ -103,45 +80,6 @@ export const ReturnErkhetConfigAddSheet = ({ onSubmit, loading }: Props) => {
                           <Form.Control>
                             <Input {...field} placeholder="User Email" />
                           </Form.Control>
-                          <Form.Message />
-                        </Form.Item>
-                      )}
-                    />
-                  </div>
-                  <div className="flex flex-col gap-4">
-                    <Form.Field
-                      control={form.control}
-                      name="boardId"
-                      render={({ field }) => (
-                        <Form.Item>
-                          <Form.Label>Board</Form.Label>
-                          <SelectBoard
-                            mode="single"
-                            value={field.value}
-                            onValueChange={(value) => {
-                              field.onChange(value as string);
-                              form.setValue('pipelineId', '');
-                              form.setValue('stageId', '');
-                            }}
-                            placeholder="Select board"
-                          />
-                          <Form.Message />
-                        </Form.Item>
-                      )}
-                    />
-                    <Form.Field
-                      control={form.control}
-                      name="stageId"
-                      render={({ field }) => (
-                        <Form.Item>
-                          <Form.Label>Stage</Form.Label>
-                          <SelectStage
-                            mode="single"
-                            value={field.value}
-                            onValueChange={(value) => field.onChange(value as string)}
-                            pipelineId={selectedPipelineId || undefined}
-                            placeholder="Select stage"
-                          />
                           <Form.Message />
                         </Form.Item>
                       )}
@@ -168,6 +106,9 @@ export const ReturnErkhetConfigAddSheet = ({ onSubmit, loading }: Props) => {
                         </Form.Item>
                       )}
                     />
+                  </div>
+                  <div className="flex flex-col gap-4">
+                    <PipelineSelectorFields form={form} />
                   </div>
                 </div>
               </div>
